// src/screens/PaymentPage/PaySection/YandexPay/yapay.types.ts
export type Price = string;

/**
 * Платежная криптограмма.
 */
export type PaymentToken = string;

/**
 * Окружение, в котором совершается платеж.
 */
export enum PaymentEnv {
  Production = 'PRODUCTION',
  Sandbox = 'SANDBOX',
}

/**
 * Код страны по стандарту ISO 3166-1 alpha-2.
 */
export enum CountryCode {
  Ru = 'RU',
  Us = 'US',
  By = 'BY',
}

/**
 * Код валюты по стандарту ISO 4217.
 */
export enum CurrencyCode {
  Rub = 'RUB',
  Byn = 'BYN',
  Usd = 'USD',
  Eur = 'EUR',
}

/**
 * Информация о продавце.
 */

export interface Merchant {
  /**
   * ID продавца.
   */
  id: string;

  /**
   * Имя продавца (отображается на платежной форме).
   */
  name: string;

  /**
   * URL продавца (необязательное поле).
   */
  url?: string;
}

/**
 * Информация о заказе.
 */

export interface OrderTotal {
  /**
   * Текст для поля total.
   */
  label?: string;
  amount: Price;
}

export interface OrderItem {
  label: string;
  amount: Price;
}

export interface Order {
  /**
   * ID заказа на стороне продавца.
   */
  id: string;
  total: OrderTotal;
  items?: OrderItem[];
}

/**
 * Описание методов оплаты.
 */

export enum PaymentMethodType {
  Card = 'CARD',
}

export enum AllowedAuthMethod {
  CloudToken = 'CLOUD_TOKEN',
  PanOnly = 'PAN_ONLY',
}

export enum AllowedCardNetwork {
  AmericanExpress = 'AMEX',
  Discover = 'DISCOVER',
  Jcb = 'JCB',
  Mastercard = 'MASTERCARD',
  Visa = 'VISA',
  VisaElectron = 'VISAELECTRON',
  Maestro = 'MAESTRO',
  Mir = 'MIR',
  UnionPay = 'UNIONPAY',
  Uzcard = 'UZCARD',
}
// ...
export type PaymentMethod = CardPaymentMethod;
// ...
export interface RequiredFields {
  billingContact?: RequiredBillingContactFields;
}

/**
 * Возвращаемые данные плательщика.
 */
export interface BillingContactField {
  // Поле появится, если было запрошено в RequiredBillingContactFields.
  email?: string;
}

/**
 * Настройки для платежа.
 */

export interface PaymentSheet {
  /**
   * Версия клиентского SDK.
   */
  version: number;

  /**
   * countryCode необходим для продавцов из Европейской экономической зоны.
   */
  countryCode: CountryCode;
  currencyCode: CurrencyCode;
  merchant: Merchant;
  order: Order;
  paymentMethods: PaymentMethod[];
  requiredFields?: RequiredFields;
}

export interface PaymentData extends PaymentSheet {
  /**
   * @default PaymentEnv.PRODUCTION
   */
  env?: PaymentEnv;
}